Make deeplinking work on ios
Update packages so we can use flutter 2.0. This seems to fix deep
linking with Router. Currently, with flutter 2.0.1, deep linking
does not work on android. It couses a crash, as described in
https://github.com/flutter/flutter/issues/75261. The fix to this
issue has been merged, but apparantly it's not included in the
release 2.0.1.
The deep links are now only with custom scheme "thaliapp://".
For universal links (https), we need to work with the website.