Fix page transition animations
Added TypedMaterialPage which adds a key based on its child's type.
That way, replacement transitions between different screen types are
animated. The LoginScreen is moved out of the navigation stack.
As logging out also removes many BlocProviders, pages can no longer be
shown after logging out, during the transition to the loginscreen.
As a solution, the login screen does not do page transitions to/from
other screens, by having it replace the navigator, instead of being in
the navigator stack.