Move handling of notification tokens to AuthCubit
The AuthCubit now takes care of anything that needs to happen upon
logging in, out, or loading the auth status. This includes setting up and
tearing down push notifications. Upon logging out, the existing fcm
token is now disabled, and marked as inactive in the backend.
Furthermore, the settingscreen now displays a warning if notification
permissions have not been granted.